requirements.txt
GitPython==3.1.18
PyYAML==5.4.1
